+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.51] - 2018-09-19
 ### Added
 - Added CompositeConfigService to allow composition of several different config sources.
 - SSMParameterStoreConfigService: Added set method.